pyproject.toml
[project]
name = "runpod-cli-dashboard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "hydra-core>=1.3.2",
    "paramiko>=4.0.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
